+++ models/contacts.js
@@ -4,17 +4,17 @@ const { randomUUID } = require("crypto");
 
 const contactsPath = path.join(__dirname, "./contacts.json");
 
-const Joi = require("joi");
-const schema = Joi.object({
-  name: Joi.string().alphanum().min(3).max(30).required(),
-  email: Joi.string()
-    .email({
-      minDomainSegments: 2,
-      tlds: { allow: ["com", "net"] },
-    })
-    .required(),
-  phone: Joi.number().required(),
-});
+// const Joi = require("joi");
+// const schema = Joi.object({
+//   name: Joi.string().alphanum().min(3).max(30).required(),
+//   email: Joi.string()
+//     .email({
+//       minDomainSegments: 2,
+//       tlds: { allow: ["com", "net"] },
+//     })
+//     .required(),
+//   phone: Joi.number().required(),
+// });
 
 const listContacts = async () => {
   const data = await fs.readFile(contactsPath);
@@ -24,34 +24,30 @@ const listContacts = async () => {
 
 const getContactById = async (contactId) => {
   const contacts = await listContacts();
-  const result = contacts.find((item) => item.id === contactId);
-  if (!result) {
+  const contact = contacts.filter((el) => el.id === contactId);
+  if (!contact) {
     return null;
   }
-  return result;
+  return contact;
 };
 
 const removeContact = async (contactId) => {
   const contacts = await listContacts();
 
-  const idx = contacts.findIndex((item) => item.id === contactId);
+  const idx = contacts.findIndex((contact) => contact.id === contactId);
   if (idx === -1) {
     return null;
   }
   const [removeById] = contacts.splice(idx, 1);
 
-  async function updatedContacts(contacts) {
-    await fs.writeFile(contactsPath, JSON.stringify(contacts));
-  }
-
-  await updatedContacts(contacts);
+  await fs.writeFile(contactsPath, JSON.stringify(contacts));
 
   return removeById;
 };
 
 const addContact = async ({ name, email, phone }) => {
   const contacts = await listContacts();
-  const newContact = schema.validate({ id: randomUUID(), name, email, phone });
+  const newContact = { id: randomUUID(), name, email, phone };
   contacts.push(newContact);
   await fs.writeFile(contactsPath, JSON.stringify(contacts));
   if (!newContact) {
@@ -63,11 +59,10 @@ const addContact = async ({ name, email, phone }) => {
 const updateContact = async (id, body) => {
   const contacts = await listContacts();
   const index = contacts.findIndex((contact) => contact.id === id);
-
   if (index === -1) {
     return null;
   }
-  contacts[index] = { ...body, id };
+  contacts[index] = { id, ...body };
   await fs.writeFile(contactsPath, JSON.stringify(contacts));
   return contacts[index];
 };
