+++ package.json
@@ -12,6 +12,7 @@
     "chalk": "^4.0",
     "cors": "2.8.5",
     "cross-env": "7.0.3",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "joi": "^17.7.0",
     "morgan": "1.10.0"
