+++ routes/api/contacts.js
@@ -8,20 +8,18 @@ const {
 
 const express = require("express");
 const router = express.Router();
+const Joi = require("joi");
 
-// const { randomUUID } = require("crypto");
-
-// const Joi = require("joi");
-// const schema = Joi.object({
-//   name: Joi.string().alphanum().min(3).max(30).required(),
-//   email: Joi.string()
-//     .email({
-//       minDomainSegments: 2,
-//       tlds: { allow: ["com", "net"] },
-//     })
-//     .required(),
-//   phone: Joi.number().required(),
-// });
+const schema = Joi.object({
+  name: Joi.string().alphanum().min(3).max(30).required(),
+  email: Joi.string()
+    .email({
+      minDomainSegments: 2,
+      tlds: { allow: ["com", "net"] },
+    })
+    .required(),
+  phone: Joi.number().required(),
+});
 
 router.get("/contacts", async (req, res, next) => {
   try {
@@ -32,30 +30,30 @@ router.get("/contacts", async (req, res, next) => {
 
 router.get("/contacts/:id", async (req, res, next) => {
   try {
-    const id = req.params.id;
+    const { id } = req.params;
     const contact = await getContactById(id);
     if (!contact) {
       return res.status(404).json({ message: "User wasn`t found" });
     }
     res.status(200).json(contact);
   } catch (error) {
-    res.status(404).json({ message: "Not found" });
+    console.log(error);
   }
 });
 
 router.post("/contacts", async (req, res, next) => {
   try {
-    const { name, email, phone } = req.body;
-    if (!name || !email || !phone) {
-      res.status(400).json({ message: "missing required name field" });
+    const { name, email, phone } = schema.validate(req.body);
+    const body = {
+      name,
+      email,
+      phone,
+    };
+    if (!body) {
+      return res.status(400).json({ message: "missing required field" });
     }
-    // const newContact = {
-    //   id: randomUUID(),
-    //   name,
-    //   email,
-    //   phone,
-    // };
-    const newContact = await addContact(name, email, phone);
+
+    const newContact = await addContact(body);
     res.status(201).json(newContact);
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -64,17 +62,8 @@ router.post("/contacts", async (req, res, next) => {
 
 router.delete("/contacts/:id", async (req, res, next) => {
   try {
-    const id = req.params.id;
-    const contacts = await listContacts();
-
-    const filteredContacts = contacts.filter(
-      (contact) => String(contact.id) !== id
-    );
-
-    if (filteredContacts.length === contacts.length) {
-      return res.status(404).json({ message: "Contact wasn`t found" });
-    }
-    await removeContact(filteredContacts);
+    const { id } = req.params;
+    await removeContact(id);
     res.status(200).json({ message: "Contact deleted" });
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -82,8 +71,9 @@ router.delete("/contacts/:id", async (req, res, next) => {
 });
 
 router.put("/contacts/:id", async (req, res, next) => {
-  const id = req.params.id;
-  const { name, email, phone } = req.body;
+  const { id } = req.params;
+  const { name, email, phone } = schema.validate(req.body);
+
   const body = {
     name,
     email,
@@ -91,12 +81,13 @@ router.put("/contacts/:id", async (req, res, next) => {
   };
 
   if (!body) {
-    res.status(400).json({ message: "missing required field" });
+    res.status(400).json({ message: "missing required fields" });
   }
   const updatedContact = await updateContact(id, body);
   if (!updatedContact) {
     return res.status(404).json({ message: "Contact wasn`t found" });
   }
+  res.status(200);
 });
 
 module.exports = router;
